Move fuzzy coordinatelist to general factories and add to registration
