+++ app/tests/annotations_tests/factories.py
@@ -17,28 +17,7 @@ from grandchallenge.annotations.models import (
     SingleLandmarkAnnotation,
 )
 from tests.cases_tests.factories import ImageFactory
-from tests.factories import UserFactory
-
-
-class FuzzyFloatCoordinatesList(factory.fuzzy.BaseFuzzyAttribute):
-    def __init__(self, size=None):
-        self.size = size
-
-    def fuzz(self):
-        if self.size is None:
-            size = factory.fuzzy.random.randgen.randint(1, 30)
-        else:
-            size = self.size
-
-        fuzzy_list = []
-        for i in range(size):
-            fuzzy_list.append(
-                [
-                    factory.fuzzy.random.randgen.uniform(0.0, 1000.0),
-                    factory.fuzzy.random.randgen.uniform(0.0, 1000.0),
-                ]
-            )
-        return fuzzy_list
+from tests.factories import UserFactory, FuzzyFloatCoordinatesList
 
 
 class DefaultImageAnnotationModelFactory(factory.DjangoModelFactory):
