+++ app/tests/registrations_tests/factories.py
@@ -2,6 +2,7 @@ import random
 import factory.fuzzy
 from grandchallenge.registrations.models import OctObsRegistration
 from tests.cases_tests.factories import ImageFactory
+from tests.factories import FuzzyFloatCoordinatesList
 
 
 class OctObsRegistrationFactory(factory.DjangoModelFactory):
@@ -11,7 +12,4 @@ class OctObsRegistrationFactory(factory.DjangoModelFactory):
     obs_image = factory.SubFactory(ImageFactory)
     oct_image = factory.SubFactory(ImageFactory)
 
-    registration_values = [
-        [random.uniform(0.0, 1000.0), random.uniform(0.0, 1000.0)],
-        [random.uniform(0.0, 1000.0), random.uniform(0.0, 1000.0)],
-    ]
+    registration_values = FuzzyFloatCoordinatesList(2)
